# Reduce calls to TotalPortfolioValue