<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- For performance, reduce the amount of call to `TotalPortfolioValue`

**Local performance tests: StatelessCoarseUniverseSelectionBenchmark 01-06-2017 to 01-01-2018**
`Master`
36.37 seconds at 31k data points per second. Processing total of 1,133,250 data points.
35.68 seconds at 32k data points per second. Processing total of 1,133,250 data points.
37.17 seconds at 30k data points per second. Processing total of 1,133,250 data points.

`PR`
28.46 seconds at 40k data points per second. Processing total of 1,133,250 data points.
29.53 seconds at 38k data points per second. Processing total of 1,133,250 data points.
30.25 seconds at 37k data points per second. Processing total of 1,133,250 data points.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Related to #3076
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Lean runs as fast as possible
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->